+++ CHANGELOG.md
@@ -18,7 +18,7 @@ to: com.cloudentity.pyron.plugin.impl.transform.request.TransformRequestPlugin
 - Update version of vertx-tools dependency to 1.9.0
 
 ### Fixed
-- Generating opeanapi
+- Generating openapi - fixed api's paths(were with regex group), fixed matching api's operations  
 
 ## [1.8.0] - 2021-05-14
 ### Added